Time: 7 ms (63.45%) | Memory: 16.6 MB (15.41%) - LeetSync

+++ 227-basic-calculator-ii/basic-calculator-ii.cpp
@@ -5,10 +5,10 @@ public:
        int n = s.size();
     
        int num = 0;
-       int res = 0;
+       long long res = 0;
        char op = '+';
 
-       stack<int> st;
+       stack<long long> st;
 
        for (int i = 0; i < n; i++)
        {
